LPD-58324 Move tests to the most appropriated test class
